% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1.R
\name{figure1}
\alias{figure1}
\title{Code that creates figure 1}
\usage{
figure1(file = "./inst/ms-figure1.pdf", col1 = "#594023",
  col2 = "#9eb844")
}
\arguments{
\item{file}{a character string giving the name of the file (see \linkpdf)}

\item{col1}{A specification for the first color.}

\item{col2}{A specification for the second color.}
}
\value{
Cretes a figure  An object of the class \code{alienPredict} is returned .
}
\description{
Format alien output and return an object of class \code{alienPredict}.
}
\examples{
figure1()

}
\author{
Nicolas Casajus & Kevin Cazelles
}
