% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1a.R
\name{figure1a}
\alias{figure1a}
\title{Code that creates figure 1a}
\usage{
figure1a(file = "ms-figure1a.pdf", col1 = "#594023", col2 = "#9eb844")
}
\arguments{
\item{file}{a character string giving the name of the file (see \code{\link[gr]{pdf}}).}

\item{col1}{A specification for the first color.}

\item{col2}{A specification for the second color.}
}
\value{
Creates a figure.
}
\description{
Code that creates figure 1a.
}
\examples{
figure1a()

}
\author{
Kevin Cazelles & Nicolas Casajus
}
