% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1b.R
\name{figure1b}
\alias{figure1b}
\title{Code that creates figure 1b found in Legagneux et al. 2018}
\usage{
figure1b(time, CC_f, PB_f, colCC = "#594023", colBD = "#9eb844")
}
\arguments{
\item{time}{time values.}

\item{CC_f}{Time series values of the research funding allocated to climate change.}

\item{PB_f}{Time series values of the research funding allocated to biodiversity.}

\item{colCC}{A specification for the first color associated with climate change data.}

\item{colBD}{A specification for the second color associated with biodiversity data.}
}
\value{
Creates a figure.
}
\description{
Code that creates a figure presenting the number of worldwide scientific
papers published in peer-reviewed literature on biodiversity (\code{colBD})
and climate change (\code{colCC}) from 1991 to 2016.
}
\examples{
# Not run:
# data(fundingUSCAN)
# grDevices::png('inst/fig/ms-figure1b.png', height = 5, width = 7, res = 300, unit = 'in')
#   figure1b(
#     fundingUSCAN$year,
#     fundingUSCAN$CC_CA + fundingUSCAN$CC_US,
#     fundingUSCAN$BD_CA + fundingUSCAN$BD_US,
#     colCC = '#b57653', colBD = '#8bcf35'
#   )
# grDevices::dev.off()

}
\author{
Kevin Cazelles & Nicolas Casajus
}
