% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2.R
\name{figure2}
\alias{figure2}
\title{Code that creates figure 2}
\usage{
figure2(pts, monthcan, monthusa, monthgbr, col1 = "#594023",
  col2 = "#9eb844")
}
\arguments{
\item{pts}{a dataframe of events.}

\item{monthcan}{monthly publications -- Canada.}

\item{monthusa}{monthly publications -- USA.}

\item{monthgbr}{monthly publications -- Great Britain.}

\item{col1}{A specification for the first color.}

\item{col2}{A specification for the second color.}
}
\value{
Creates figure 2.
}
\description{
Code that creates figure 2.
}
\examples{
#-- Not run
# data(events); data(monthcan); data(monthusa); data(monthgbr)
# grDevices::png(file='inst/fig/ms-figure2.png', width = 8, height = 6, res = 300, unit = 'in')
#  figure2(events, monthcan, monthusa, monthgbr)
# grDevices::dev.off()

}
\author{
Nicolas Casajus & Kevin Cazelles
}
