% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fundingCAN_BD}
\alias{fundingCAN_BD}
\title{Canadian fundings allocated to Biodiversity research.}
\format{A data frame of 29 rows and 15 columns}
\usage{
data(fundingCAN_BD)
}
\description{
Canadian fundings allocated to Biodiversity research.
}
\author{
tobeadded
}
\keyword{datasets}
