% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fundingCAN_CC}
\alias{fundingCAN_CC}
\title{Canadian fundings allocated to Climate Change research.}
\format{A data frame of 29 rows and 15 columns}
\usage{
data(fundingCAN_CC)
}
\description{
Canadian fundings allocated to Climate Change research.
}
\author{
tobeadded
}
\keyword{datasets}
