% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fundingUK}
\alias{fundingUK}
\title{British fundings allocated to Biodiversity and Climate Change research.}
\format{A data frame of 29 rows and 15 columns}
\usage{
data(fundingUK)
}
\description{
British fundings allocated to Climate Change research.
}
\author{
tobeadded
}
\keyword{datasets}
