% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fundingUSCAN}
\alias{fundingUSCAN}
\title{Canadian and American fundings allocated to Climate Change and BioDiversity research.}
\format{A data frame of 25 rows and 6 columns}
\usage{
data(fundingUSCAN)
}
\description{
Canadian and American fundings allocated to Climate Change and BioDiversity research.
}
\author{
tobeadded
}
\keyword{datasets}
