% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthcan}
\alias{monthcan}
\title{Monthly record of publicatoin in Canada for cc and bd}
\format{A data frame of 324 rows and 3 columns}
\usage{
data(monthcan)
}
\description{
Monthly record of publicatoin in Canada for cc and bd
}
\author{
tobeadded
}
\keyword{datasets}
