% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthgbr}
\alias{monthgbr}
\title{Monthly record of publicatoin in UK for cc and bd}
\format{A data frame of 324 rows and 3 columns}
\usage{
data(monthgbr)
}
\description{
Monthly record of publicatoin in UK for cc and bd
}
\author{
tobeadded
}
\keyword{datasets}
