% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newsNames}
\alias{newsNames}
\title{Newspapers' names.}
\format{A data frame of 12 rows (1 per journal) and 2 columns: \code{country}
and \code{name}.}
\usage{
data(newsNames)
}
\description{
Names of the newspapers included in the study.
}
\author{
Kevin Cazelles
}
\keyword{datasets}
