% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newsPapers}
\alias{newsPapers}
\title{Monthly record of newspapers articles dealing with climate change (CC) or
biodiversity (BD).}
\format{A data frame of 324 rows and 3 columns}
\usage{
data(newsPapers)
}
\description{
Monthly record of newspapers articles dealing with climate change (CC) and/or
biodiversity (BD). Articles recorded were extracted from newspapers we assumed
to be the most popular in USA, Canada and UK.
}
\author{
tobeadded
}
\keyword{datasets}
