% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rateExch}
\alias{rateExch}
\title{Exchange rates}
\format{A data frame of 27 rows and 5 columns}
\usage{
data(rateExch)
}
\description{
Exchange rates (USD-POUND-CAND) from 1990 to 2016. All values are
exchanges rates averaged over a year.
}
\author{
tobeadded
}
\keyword{datasets}
