% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sciNames}
\alias{sciNames}
\title{Number of scientific papers retrieved per journal title.}
\format{A data frame of 10923 rows and 3 columns: \code{sourcesTitles} giving
journal titles, \code{records} that gives the number of papers we retrieved
from a specific title and \code{keyword} that indicates the keyword (CC stands
for climate change and BD for biodiversity) to which the number of articles
refers to.}
\usage{
data(sciNames)
}
\description{
This dataset includes the total of scientific papers we retrieved from Web of
Science on March 3rd, 2017 per journal title.
}
\author{
Kevin Cazelles and Nicolas Casajus
}
\keyword{datasets}
