% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sciPapers}
\alias{sciPapers}
\title{Worldwide scientific papers published in peer-reviewed literature on
biodiversity and climate change.}
\format{A data frame of 26 rows and 3 columns: \code{year}, \code{records_BD}
and \code{records_CC}.}
\usage{
data(sciPapers)
}
\description{
Worldwide scientific papers published in peer-reviewed literature dealing with
biodiversity (BD) or climate change (CC) retrieved from Web of Science on
March 3rd, 2017.
}
\author{
Steve Vissault and Nicolas Casajus
}
\keyword{datasets}
