% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPeaks.R
\name{detectPeaks}
\alias{detectPeaks}
\title{Detect peaks in the newspapers articles time series.}
\usage{
detectPeaks(newsPapers)
}
\arguments{
\item{newsPapers}{A data frame that includes the number of newpapers articles
for climate change and}
}
\value{
A list of two elements: significant peaks fot the two
}
\description{
Detect peaks in the newspapers articles time series.
}
\examples{
data(newsPapers)
newsPapers
detectPeaks(newsPapers)
}
\author{
Nicolas Casajus & Kevin Cazelles
}
