% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events}
\alias{events}
\title{Major events}
\format{A  data frame of 29 rows and 15 columns}
\usage{
data(events)
}
\description{
A currantion of important international event/conference related to either
biodiversity or climate change.
}
\author{
tobeadded
}
\keyword{datasets}
