% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1a.R
\name{figure1a}
\alias{figure1a}
\title{Code that creates figure 1a}
\usage{
figure1a(time, CC_f, PB_f, col1 = "#594023", col2 = "#9eb844")
}
\arguments{
\item{time}{time values.}

\item{CC_f}{Time-series values of the research funding allocated to climate change.}

\item{PB_f}{Time-series values of the research funding allocated to biodiversity.}

\item{col1}{A specification for the first color.}

\item{col2}{A specification for the second color.}
}
\value{
Creates a figure.
}
\description{
Code that creates figure 1a.
}
\examples{
# Not run:
# data(fundingUSCAN)
# grDevices::png('inst/fig/ms-figure1a.png', height = 5, width = 7, res = 300, unit = 'in')
#   figure1a(
#     fundingUSCAN$year,
#     fundingUSCAN$CC_CA + fundingUSCAN$CC_US,
#     fundingUSCAN$PB_CA + fundingUSCAN$PB_US
#   )
# grDevices::dev.off()

}
\author{
Kevin Cazelles & Nicolas Casajus
}
