% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1a.R
\name{figure1a}
\alias{figure1a}
\title{Code that creates figure 1a}
\usage{
figure1a(time, CC_p, PB_p, colCC = "#594023", colBD = "#9eb844")
}
\arguments{
\item{time}{time values.}

\item{CC_p}{Time-series values of the research papers pertaining to climate change.}

\item{PB_p}{Time-series values of the research papers pertaining to biodiversity.}

\item{colCC}{A specification for the first color associated with climate change data.}

\item{colBD}{A specification for the second color associated with biodiversity data.}
}
\value{
Creates a figure.
}
\description{
Code that creates figure 1a found in Legagneux et al. 2018.
}
\examples{
# Not run:
# data(sciPapers)
# grDevices::png('inst/fig/ms-figure1a.png', height = 5, width = 7, res = 300, unit = 'in')
#   figure1a(
#     sciPapers$year,
#     sciPapers$records_CC,
#     sciPapers$records_BD
#   )
# grDevices::dev.off()

}
\author{
Kevin Cazelles & Nicolas Casajus
}
