% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1b.R
\name{figure1b}
\alias{figure1b}
\title{Code that creates figure 1b}
\usage{
figure1b(time, CC_f, PB_f, col1 = "#594023", col2 = "#9eb844")
}
\arguments{
\item{time}{time values.}

\item{CC_f}{Time-series values of the research funding allocated to climate change.}

\item{PB_f}{Time-series values of the research funding allocated to biodiversity.}

\item{col1}{A specification for the first color.}

\item{col2}{A specification for the second color.}
}
\value{
Creates a figure.
}
\description{
Code that creates a figure presenting the number of worldwide scientific
papers published in peer-reviewed literature on biodiversity (\code{col2})
and climate change (\code{col1}) from 1991 to 2016.
}
\examples{
# Not run:
# data(fundingUSCAN)
# grDevices::png('inst/fig/ms-figure1b.png', height = 5, width = 7, res = 300, unit = 'in')
#   figure1b(
#     fundingUSCAN$year,
#     fundingUSCAN$CC_CA + fundingUSCAN$CC_US,
#     fundingUSCAN$BD_CA + fundingUSCAN$BD_US,
#     col1 = '#b57653', col2 = '#8bcf35'
#   )
# grDevices::dev.off()

}
\author{
Kevin Cazelles & Nicolas Casajus
}
