% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2.R
\name{figure2}
\alias{figure2}
\title{Code that creates figure 2}
\usage{
figure2(pts, monthcan, monthusa, monthgbr, file = "ms-figure2.pdf",
  col1 = "#594023", col2 = "#9eb844")
}
\arguments{
\item{pts}{a dataframe of events.}

\item{monthcan}{monthly publications -- Canada.}

\item{monthusa}{monthly publications -- USA.}

\item{monthgbr}{monthly publications -- Great Britain.}

\item{file}{a character string giving the name of the file (see \code{\link[grDevices]{pdf}}).}

\item{col1}{A specification for the first color.}

\item{col2}{A specification for the second color.}
}
\value{
Creates a figure 2 as a pdf file.
}
\description{
Code that creates figure 2.
}
\examples{
#-- Not run
# data(events); data(monthcan); data(monthusa); data(monthgbr)
# figure2(events, monthcan, monthusa, monthgbr)

}
\author{
Nicolas Casajus & Kevin Cazelles
}
