% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2.R
\name{figure2}
\alias{figure2}
\title{Code that creates figure 2}
\usage{
figure2(file = "ms-figure2.pdf", col1 = "#594023", col2 = "#9eb844")
}
\arguments{
\item{file}{a character string giving the name of the file (see \code{\link[gr]{pdf}}).}

\item{col1}{A specification for the first color.}

\item{col2}{A specification for the second color.}
}
\value{
Creates a figure.
}
\description{
Format alien output and return an object of class \code{alienPredict}.
}
\examples{
figure2()

}
\author{
Nicolas Casajus & Kevin Cazelles
}
