% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2.R
\name{figure2}
\alias{figure2}
\title{Code that creates figure 2 found in Legagneux et al. (2018)}
\usage{
figure2(pts, newsPapers, colCC = "#594023", colBD = "#9eb844")
}
\arguments{
\item{pts}{A dataframe of events.}

\item{newsPapers}{Monthly publications -- Canada + USA + UK.}

\item{colCC}{A specification for the first color associated with climate change data.}

\item{colBD}{A specification for the second color associated with biodiversity data.}
}
\value{
Creates figure 2.
}
\description{
Code that creates figure 2, that is the number of newspaper
articles published per month on biodiversity (\code{colBD}) or climate
change (\code{colCC}) issues in US, Canada and UK. Detected peaks (plain
dots) and associated events are shown. Peaks without associated events (empty
dots) could not be associated with a priori events. Events that embraced both
CC and BD issues are written in bold.
}
\examples{
#-- Not run:
# data(events); data(newsPapers);
# grDevices::png(file='inst/fig/ms-figure2.png', width = 8, height = 6, res = 300, unit = 'in')
#  figure2(events, newsPapers)
# grDevices::dev.off()

}
\author{
Nicolas Casajus & Kevin Cazelles
}
