% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newsPapers}
\alias{newsPapers}
\title{Monthly record of newspapers articles dealing with climate change (CC) or
biodiversity (BD).}
\format{A data frame of 324 rows and 7 columns: column \code{month} records
the month associated with the number of articles recorded in the 6 other
columns, that is articles pertaining to either climate change (CC)
and/or biodiversity (BD) found in the most popular journals in the three
countries abovementioned.}
\usage{
data(newsPapers)
}
\description{
Time series of monthly record of newspapers articles dealing with climate
change (CC) and/or biodiversity (BD). Articles recorded were extracted from
newspapers we assumed to be the most popular in Canada (CA), United States
of America (US) and the United Kingdom (UK) (see \code{titles} dataset for
more details).
}
\author{
Nicolas Casajus and Pierre Legagneux
}
\keyword{datasets}
